Implemented add variable add action. Originally developed by dawehner.
